+++ src/oidc_provider_mock/_app.py
@@ -4,14 +4,13 @@ import textwrap
 import warnings
 from collections.abc import Callable, Sequence
 from dataclasses import dataclass
-from datetime import datetime, timedelta, timezone
+from datetime import UTC, datetime, timedelta
 from http import HTTPStatus
-from typing import TypeVar, cast
+from typing import Never, cast, override
 from urllib.parse import parse_qs, urlencode, urlparse, urlunparse
 from uuid import uuid4
 
 import authlib.deprecate
-import authlib.integrations.flask_oauth2 as flask_oauth2
 import authlib.oauth2.rfc6749
 import authlib.oauth2.rfc6749.errors
 import authlib.oauth2.rfc6750
@@ -23,9 +22,9 @@ import werkzeug.debug
 import werkzeug.exceptions
 import werkzeug.local
 from authlib import jose
+from authlib.integrations import flask_oauth2
 from authlib.integrations.flask_oauth2.requests import FlaskOAuth2Request
 from authlib.oauth2 import OAuth2Error, OAuth2Request
-from typing_extensions import Never, override
 
 from . import _client
 from ._storage import (
@@ -188,6 +187,17 @@ authorization = cast(
 blueprint = flask.Blueprint("oidc-provider-mock", __name__)
 
 
+@blueprint.after_request
+def add_cors_headers(response: flask.Response) -> flask.Response:
+    if flask.request.endpoint == f"{blueprint.name}.{authorize.__name__}":
+        return response
+
+    response.headers["Access-Control-Allow-Origin"] = "*"
+    response.headers["Access-Control-Allow-Headers"] = "*, Authorization"
+    response.headers["Access-Control-Allow-Methods"] = "GET, POST, PUT, OPTIONS"
+    return response
+
+
 @dataclass(kw_only=True, frozen=True)
 class Config:
     require_client_registration: bool = False
@@ -251,8 +261,7 @@ def setup(setup_state: flask.blueprints.BlueprintSetupState):
                 user_id=request.user.sub,
                 # request.scope may actually be None
                 scope=scope,
-                expires_at=datetime.now(timezone.utc)
-                + timedelta(seconds=token["expires_in"]),
+                expires_at=datetime.now(UTC) + timedelta(seconds=token["expires_in"]),
             )
         )
 
@@ -266,7 +275,7 @@ def setup(setup_state: flask.blueprints.BlueprintSetupState):
                     token=token["refresh_token"],
                     user_id=request.user.sub,
                     scope=scope,
-                    expires_at=datetime.now(timezone.utc)
+                    expires_at=datetime.now(UTC)
                     + timedelta(seconds=token["expires_in"]),
                     client_id=request.client.id,
                 )
@@ -455,12 +464,14 @@ def authorize() -> flask.typing.ResponseReturnValue:
     recent_subjects = [
         sub for sub in storage.get_recent_subjects() if sub not in predefined_users
     ]
+    scopes = flask.request.args.get("scope", "").split()
 
     if flask.request.method == "GET":
         return flask.render_template(
             "authorization_form.html",
             redirect_uri=redirect_uri,
             client_id=grant.client.id,
+            scopes=scopes,
             recent_subjects=recent_subjects,
             predefined_users=predefined_users,
         )
@@ -473,14 +484,10 @@ def authorize() -> flask.typing.ResponseReturnValue:
             )
 
         sub = flask.request.form.get("sub")
-        if sub is None:
-            return flask.render_template(
-                "authorization_form.html",
-                redirect_uri=redirect_uri,
-                client_id=grant.client.id,
-                recent_subjects=recent_subjects,
-                predefined_users=predefined_users,
-                sub_missing=True,
+        if not sub:
+            raise _AuthorizationValidationException(
+                authlib.oauth2.rfc6749.InvalidRequestError.error,
+                "Missing 'sub' form parameter",
             )
 
         user = storage.get_user(sub)
@@ -686,10 +693,10 @@ blueprint.register_error_handler(
     _insecure_transport_error_handler,
 )
 
-_Model = TypeVar("_Model", bound=pydantic.BaseModel)
 
-
-def _validate_body(request: flask.Request, model: type[_Model]) -> _Model:
+def _validate_body[Model: pydantic.BaseModel](
+    request: flask.Request, model: type[Model]
+) -> Model:
     try:
         return model.model_validate(request.json, strict=True)
     except pydantic.ValidationError as error: