+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 
 ## Unreleased
 
-- Added --user-claims-file to CLI which allows loading user claims as a YAML/JSON file.
+- Added `--user-claims-file` to CLI which allows loading user claims as a YAML/JSON file.
+
+## v0.4.1 - 2026-04-12
+
+- Add CORS headers ([@tjarbo][])
+
+## v0.4.0 - 2026-03-27
+
+- Update authorization UI ([@Kenneth-Sills][])
+- Require Python >=3.12 and Pydantic >=2.3
+
+[@Kenneth-Sills]: https://github.com/Kenneth-Sills
 
 ## v0.3.4 - 2026-01-25